+++ CHSZiK/src/app/main/pages/list-of-equipment/components/modal-position/modal-position.component.ts
@@ -2,9 +2,10 @@ import {Component, Inject, OnInit} from '@angular/core';
 import {FormControl, FormGroup, Validators} from "@angular/forms";
 import {ListOfLocationService} from "../../../list-of-location/services/list-of-location.service";
 import {ListOfEquipmentService} from "../../services/list-of-equipment.service";
-import {MAT_DIALOG_DATA, MatDialogRef} from "@angular/material/dialog";
+import {MAT_DIALOG_DATA, MatDialog, MatDialogRef} from "@angular/material/dialog";
 import {IPosition} from "../../../../../models/position";
 import {ToastrService} from "ngx-toastr";
+import {ModalConfirmationComponent} from "../../../modal-confirmation/modal-confirmation.component";
 
 interface IModelsData {
   modelsId: number
@@ -16,7 +17,7 @@ interface IModelsData {
   styleUrls: ['./modal-position.component.scss']
 })
 export class ModalPositionComponent implements OnInit {
-  constructor(private listOfLocationService: ListOfLocationService, private listOfEquipmentService: ListOfEquipmentService, public dialogRef: MatDialogRef<ModalPositionComponent>, @Inject(MAT_DIALOG_DATA) public data: IModelsData, private toastrService: ToastrService) {
+  constructor(private listOfLocationService: ListOfLocationService, private dialog: MatDialog, private listOfEquipmentService: ListOfEquipmentService, public dialogRef: MatDialogRef<ModalPositionComponent>, @Inject(MAT_DIALOG_DATA) public data: IModelsData, private toastrService: ToastrService) {
   }
   positionsArr: IPosition[] = []
   addingPositionForm!: FormGroup;
@@ -26,10 +27,19 @@ export class ModalPositionComponent implements OnInit {
   }
 
   deletePosition(group_id: string) {
-    this.listOfEquipmentService.deletePosition(group_id).subscribe(() => {
-      this.toastrService.success('Позиция удалена')
-      this.openPositions(this.data.modelsId)
+    this.dialog.open(ModalConfirmationComponent).componentInstance.sendConfirmation.subscribe((send) => {
+      if(send) {
+        this.listOfEquipmentService.deletePosition(group_id).subscribe(() => {
+          this.toastrService.success('Позиция удалена')
+          this.openPositions(this.data.modelsId)
+        })
+      }
     })
+
+    // this.listOfEquipmentService.deletePosition(group_id).subscribe(() => {
+    //   this.toastrService.success('Позиция удалена')
+    //   this.openPositions(this.data.modelsId)
+    // })
   }
 
   addingNewPosition() {
